3sem/misppi/lab7/python-converter/requirements.txt
anyio==3.6.2
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.87.0
h11==0.14.0
idna==3.4
lxml==4.9.1
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
pydantic==1.10.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.19.0
